Safari-609.1.11.1/LayoutTests/fast/dom/Node/appendChild-no-op-mutationobserver-expected.txt

Tests that calls of appendChild() lead to mutation events even if they are no-ops.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Mutation observer was notified
PASS wasMutationEventListenerCalled is true
PASS successfullyParsed is true

TEST COMPLETE
